refactor: streamline service imports and improve test structure
