+++ src/services/index.ts
@@ -38,3 +38,4 @@ export * from './zip.service';
 export { default as vpnAuthService } from './vpnAuth.service';
 export { default as authService } from './auth.service';
 export { default as userService } from './user.service';
+export { default as oauthService } from './oauth.service';
