+++ src/views/Checkout/services/payment.service.test.ts
@@ -2,8 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { loadStripe } from '@stripe/stripe-js/pure';
 import paymentService from './payment.service';
 import { SdkFactory } from '../../../app/core/factory/sdk';
-import envService from '../../../app/core/services/env.service';
-import localStorageService from '../../../app/core/services/local-storage.service';
+import { envService, localStorageService } from 'services';
 import axios from 'axios';
 import { UserType } from '@internxt/sdk/dist/drive/payments/types/types';
 import { StripeSessionMode } from '../types';
