+++ vitest.shared.ts
@@ -15,8 +15,6 @@ export default {
       services: path.resolve(__dirname, './src/services'),
       utils: path.resolve(__dirname, './src/utils'),
       views: path.resolve(__dirname, './src/views'),
-      services: path.resolve(__dirname, './src/services'),
-      components: path.resolve(__dirname, './src/components'),
       crypto: 'crypto-browserify',
       stream: 'stream-browserify',
       path: 'path-browserify',
